requirements.txt
streamlit
requests
Pillow
torch 
diffusers 
transformers 
fastapi 
uvicorn 
nest_asyncio
accelerate
pyngrok